+++ src/PostgraphileNestedTypesPlugin.js
@@ -16,10 +16,8 @@ module.exports = function PostGraphileNestedTypesPlugin(
             tags: { name: tagName },
           },
           isForward,
-        } = options;
-        return inflection.upperCamelCase(
-          `${tagName || name}_${isForward ? '' : 'Inverse'}_input`,
-        );
+        } = options
+        return inflection.upperCamelCase(`${tagName || name}_${isForward ? '' : 'Inverse'}_input`)
       },
       nestedCreateInputType(options) {
         const {
@@ -28,16 +26,14 @@ module.exports = function PostGraphileNestedTypesPlugin(
             tags: { name: tagName },
           },
           foreignTable,
-        } = options;
-        return inflection.upperCamelCase(
-          `${tagName || name}_${foreignTable.name}_create_input`,
-        );
+        } = options
+        return inflection.upperCamelCase(`${tagName || name}_${foreignTable.name}_create_input`)
       },
     }),
-  );
+  )
 
   builder.hook('build', (build) => {
-    const { extend, pgOmit: omit, inflection } = build;
+    const { extend, pgOmit: omit, inflection } = build
 
     return extend(build, {
       pgNestedPluginForwardInputTypes: {},
@@ -48,34 +44,29 @@ module.exports = function PostGraphileNestedTypesPlugin(
           constraint: {
             keyAttributes: keys,
             foreignKeyAttributes: foreignKeys,
-            tags: {
-              fieldName,
-              foreignFieldName,
-              forwardMutationName,
-              reverseMutationName,
-            },
+            tags: { fieldName, foreignFieldName, forwardMutationName, reverseMutationName },
           },
           table,
           isForward,
           foreignTable,
-        } = options;
-        const tableFieldName = inflection.tableFieldName(foreignTable);
-        const keyNames = keys.map((k) => k.name);
-        const foreignKeyNames = foreignKeys.map((k) => k.name);
+        } = options
+        const tableFieldName = inflection.tableFieldName(foreignTable)
+        const keyNames = keys.map((k) => k.name)
+        const foreignKeyNames = foreignKeys.map((k) => k.name)
 
         const constraints = foreignTable.constraints
           .filter((con) => con.type === 'f')
           .filter((con) => con.foreignClass.id === table.id)
-          .filter((con) => !omit(con, 'read'));
+          .filter((con) => !omit(con, 'read'))
 
-        const multipleFKs = constraints.length > 1;
+        const multipleFKs = constraints.length > 1
 
         const isUnique = !!foreignTable.constraints.find(
           (c) =>
             (c.type === 'p' || c.type === 'u') &&
             c.keyAttributeNums.length === keys.length &&
             c.keyAttributeNums.every((n, i) => keys[i].num === n),
-        );
+        )
 
         const computedReverseMutationName = inflection.camelCase(
           `${
@@ -85,48 +76,44 @@ module.exports = function PostGraphileNestedTypesPlugin(
                 : tableFieldName
               : inflection.pluralize(tableFieldName)
           }`,
-        );
+        )
 
         if (isForward) {
           if (forwardMutationName) {
-            return forwardMutationName;
+            return forwardMutationName
           }
           if (fieldName) {
-            return fieldName;
+            return fieldName
           }
           if (nestedMutationsSimpleFieldNames && !multipleFKs) {
-            return inflection.camelCase(`${tableFieldName}`);
+            return inflection.camelCase(`${tableFieldName}`)
           }
-          return inflection.camelCase(
-            `${tableFieldName}_to_${keyNames.join('_and_')}`,
-          );
+          return inflection.camelCase(`${tableFieldName}_to_${keyNames.join('_and_')}`)
         }
 
         // reverse mutation
         if (reverseMutationName) {
-          return reverseMutationName;
+          return reverseMutationName
         }
         if (foreignFieldName) {
-          return foreignFieldName;
+          return foreignFieldName
         }
         if (!multipleFKs) {
           return nestedMutationsSimpleFieldNames
             ? computedReverseMutationName
             : inflection.camelCase(
-                `${computedReverseMutationName}_using_${foreignKeyNames.join(
-                  '_and_',
-                )}`,
-              );
+                `${computedReverseMutationName}_using_${foreignKeyNames.join('_and_')}`,
+              )
         }
         // tables have mutliple relations between them
         return inflection.camelCase(
-          `${computedReverseMutationName}_to_${keyNames.join(
+          `${computedReverseMutationName}_to_${keyNames.join('_and_')}_using_${foreignKeyNames.join(
             '_and_',
-          )}_using_${foreignKeyNames.join('_and_')}`,
-        );
+          )}`,
+        )
       },
-    });
-  });
+    })
+  })
 
   builder.hook('GraphQLInputObjectType:fields', (fields, build, context) => {
     const {
@@ -141,60 +128,48 @@ module.exports = function PostGraphileNestedTypesPlugin(
       pgNestedTableDeleterFields,
       pgNestedTableUpdaterFields,
       pgNestedFieldName,
-      graphql: {
-        GraphQLInputObjectType,
-        GraphQLList,
-        GraphQLNonNull,
-        GraphQLBoolean,
-      },
-    } = build;
+      graphql: { GraphQLInputObjectType, GraphQLList, GraphQLNonNull, GraphQLBoolean },
+    } = build
 
     const {
       scope: { isInputType, isPgRowType, pgIntrospection: table },
       GraphQLInputObjectType: gqlType,
-    } = context;
+    } = context
 
     if (!isInputType || !isPgRowType) {
-      return fields;
+      return fields
     }
 
     /**
      * whitelist
      */
-    if (
-      nestedMutationsTableNameWhiteList &&
-      !nestedMutationsTableNameWhiteList[table.name]
-    )
-      return fields;
+    if (nestedMutationsTableNameWhiteList && !nestedMutationsTableNameWhiteList[table.name])
+      return fields
 
     const foreignKeyConstraints = introspectionResultsByKind.constraint
       .filter((con) => con.type === 'f')
-      .filter(
-        (con) => con.classId === table.id || con.foreignClassId === table.id,
-      )
-      .filter((con) => !omit(con, 'read'));
+      .filter((con) => con.classId === table.id || con.foreignClassId === table.id)
+      .filter((con) => !omit(con, 'read'))
 
     if (!foreignKeyConstraints.length) {
       // table has no foreign relations
-      return fields;
+      return fields
     }
 
-    const tableTypeName = gqlType.name;
+    const tableTypeName = gqlType.name
     // console.log({ tableTypeName, tableName: table.name })
-    pgNestedPluginForwardInputTypes[table.id] = [];
-    pgNestedPluginReverseInputTypes[table.id] = [];
+    pgNestedPluginForwardInputTypes[table.id] = []
+    pgNestedPluginReverseInputTypes[table.id] = []
 
     foreignKeyConstraints.forEach((constraint) => {
-      const isForward = constraint.classId === table.id;
+      const isForward = constraint.classId === table.id
       const foreignTable = isForward
         ? introspectionResultsByKind.classById[constraint.foreignClassId]
-        : introspectionResultsByKind.classById[constraint.classId];
+        : introspectionResultsByKind.classById[constraint.classId]
 
       // istanbul ignore next
       if (!foreignTable) {
-        throw new Error(
-          `Could not find the foreign table (constraint: ${constraint.name})`,
-        );
+        throw new Error(`Could not find the foreign table (constraint: ${constraint.name})`)
       }
 
       /**
@@ -205,28 +180,27 @@ module.exports = function PostGraphileNestedTypesPlugin(
         nestedMutationsTableNameWhiteList[table.name] &&
         !nestedMutationsTableNameWhiteList[table.name][foreignTable.name]
       )
-        return;
+        return
 
-      const foreignTableName = inflection.tableFieldName(foreignTable);
+      const foreignTableName = inflection.tableFieldName(foreignTable)
       // console.debug({ foreignTableName, tableName: table.name })
       // console.debug({ foreignTableName: foreignTable.name, tableName: table.name })
 
       const foreignUniqueConstraints = foreignTable.constraints
         .filter((con) => con.type === 'u' || con.type === 'p')
-        .filter((con) => !con.keyAttributes.some((key) => omit(key)));
+        .filter((con) => !con.keyAttributes.some((key) => omit(key)))
 
-      const connectable = !!foreignUniqueConstraints.length;
+      const connectable = !!foreignUniqueConstraints.length
       const creatable =
         !omit(foreignTable, 'create') &&
         !omit(constraint, 'create') &&
-        !constraint.keyAttributes.some((key) => omit(key, 'create'));
-      const updateable =
-        !omit(foreignTable, 'update') && !omit(constraint, 'update');
+        !constraint.keyAttributes.some((key) => omit(key, 'create'))
+      const updateable = !omit(foreignTable, 'update') && !omit(constraint, 'update')
       const deleteable =
         nestedMutationsDeleteOthers &&
         foreignTable.primaryKeyConstraint &&
         !omit(foreignTable, 'delete') &&
-        !omit(constraint, 'delete');
+        !omit(constraint, 'delete')
 
       if (
         (!connectable && !creatable && !deleteable && !updateable) ||
@@ -234,37 +208,37 @@ module.exports = function PostGraphileNestedTypesPlugin(
         // || primaryKey.keyAttributes.some(key => omit(key, 'read'))
         // || foreignPrimaryKey.keyAttributes.some(key => omit(key, 'read'))
       ) {
-        return;
+        return
       }
 
-      const keys = constraint.keyAttributes;
+      const keys = constraint.keyAttributes
       const isUnique = !!foreignTable.constraints.find(
         (c) =>
           (c.type === 'p' || c.type === 'u') &&
           c.keyAttributeNums.length === keys.length &&
           c.keyAttributeNums.every((n, i) => keys[i].num === n),
-      );
+      )
 
       const fieldName = pgNestedFieldName({
         constraint,
         table,
         foreignTable,
         isForward,
-      });
+      })
 
       const createInputTypeName = inflection.nestedCreateInputType({
         constraint,
         table,
         foreignTable,
         isForward,
-      });
+      })
 
       const connectorTypeName = inflection.nestedConnectorType({
         constraint,
         table,
         foreignTable,
         isForward,
-      });
+      })
 
       const connectorInputField = newWithHooks(
         GraphQLInputObjectType,
@@ -275,14 +249,14 @@ module.exports = function PostGraphileNestedTypesPlugin(
             const gqlForeignTableType = getGqlInputTypeByTypeIdAndModifier(
               foreignTable.type.id,
               null,
-            );
-            const operations = {};
+            )
+            const operations = {}
 
             if (!isForward && deleteable) {
               operations.deleteOthers = {
                 description: `Flag indicating whether all other \`${foreignTableName}\` records that match this relationship should be removed.`,
                 type: GraphQLBoolean,
-              };
+              }
             }
             // console.debug(`\n\n\n\n\n`)
             // console.debug(`${foreignTable.type}: ${foreignTable.name}`)
@@ -300,9 +274,9 @@ module.exports = function PostGraphileNestedTypesPlugin(
                       : isUnique
                       ? field
                       : new GraphQLList(new GraphQLNonNull(field)),
-                  };
+                  }
                 },
-              );
+              )
             if (deleteable) {
               pgNestedTableDeleterFields[foreignTable.id].forEach(
                 ({ field, fieldName: deleterFieldName }) => {
@@ -313,9 +287,9 @@ module.exports = function PostGraphileNestedTypesPlugin(
                       : isUnique
                       ? field
                       : new GraphQLList(new GraphQLNonNull(field)),
-                  };
+                  }
                 },
-              );
+              )
             }
             // console.debug(`\n\n\n\n\n`)
             // // console.debug(`${table.type}: ${table.name}`)
@@ -333,9 +307,9 @@ module.exports = function PostGraphileNestedTypesPlugin(
                       : isUnique
                       ? field
                       : new GraphQLList(new GraphQLNonNull(field)),
-                  };
+                  }
                 },
-              );
+              )
             }
             if (creatable) {
               const createInputType = newWithHooks(
@@ -344,14 +318,12 @@ module.exports = function PostGraphileNestedTypesPlugin(
                   name: createInputTypeName,
                   description: `The \`${foreignTableName}\` to be created by this mutation.`,
                   fields: () => {
-                    const inputFields = gqlForeignTableType._fields;
-                    const omittedFields = constraint.keyAttributes.map((k) =>
-                      inflection.column(k),
-                    );
+                    const inputFields = gqlForeignTableType._fields
+                    const omittedFields = constraint.keyAttributes.map((k) => inflection.column(k))
                     return Object.keys(inputFields)
                       .filter((key) => !omittedFields.includes(key))
                       .map((k) => Object.assign({}, { [k]: inputFields[k] }))
-                      .reduce((res, o) => Object.assign(res, o), {});
+                      .reduce((res, o) => Object.assign(res, o), {})
                   },
                 },
                 {
@@ -361,16 +333,16 @@ module.exports = function PostGraphileNestedTypesPlugin(
                   pgInflection: table,
                   pgNestedForeignInflection: foreignTable,
                 },
-              );
+              )
 
               operations.create = {
                 description: `A \`${gqlForeignTableType.name}\` object that will be created and connected to this object.`,
                 type: isForward
                   ? createInputType
                   : new GraphQLList(new GraphQLNonNull(createInputType)),
-              };
+              }
             }
-            return operations;
+            return operations
           },
         },
         {
@@ -379,7 +351,7 @@ module.exports = function PostGraphileNestedTypesPlugin(
           pgInflection: table,
           pgNestedForeignInflection: foreignTable,
         },
-      );
+      )
 
       if (isForward) {
         pgNestedPluginForwardInputTypes[table.id].push({
@@ -391,7 +363,7 @@ module.exports = function PostGraphileNestedTypesPlugin(
           foreignKeys: constraint.foreignKeyAttributes,
           connectorInputField,
           isUnique,
-        });
+        })
       } else {
         pgNestedPluginReverseInputTypes[table.id].push({
           name: fieldName,
@@ -402,10 +374,10 @@ module.exports = function PostGraphileNestedTypesPlugin(
           foreignKeys: constraint.foreignKeyAttributes,
           connectorInputField,
           isUnique,
-        });
+        })
       }
-    });
+    })
 
-    return fields;
-  });
-};
+    return fields
+  })
+}