+++ src/PostgraphileNestedUpdatersPlugin.js
@@ -5,52 +5,50 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
   builder.hook('inflection', (inflection, build) =>
     build.extend(inflection, {
       nestedUpdateByNodeIdField() {
-        return this.camelCase(`update_by_${build.nodeIdFieldName}`);
+        return this.camelCase(`update_by_${build.nodeIdFieldName}`)
       },
       nestedUpdateByKeyField(options) {
-        const { constraint } = options;
+        const { constraint } = options
         return this.camelCase(
-          `update_by_${constraint.keyAttributes
-            .map((k) => k.name)
-            .join('_and_')}`,
-        );
+          `update_by_${constraint.keyAttributes.map((k) => k.name).join('_and_')}`,
+        )
       },
       nestedUpdateByNodeIdInputType(options) {
-        const { table, constraint } = options;
+        const { table, constraint } = options
 
-        const tableFieldName = this.tableFieldName(table);
-        const parentTableFieldName = this.tableFieldName(constraint.class);
-        const constraintName = constraint.tags.name || constraint.name;
+        const tableFieldName = this.tableFieldName(table)
+        const parentTableFieldName = this.tableFieldName(constraint.class)
+        const constraintName = constraint.tags.name || constraint.name
 
         return this.upperCamelCase(
           `${tableFieldName}_on_${parentTableFieldName}_for_${constraintName}_node_id_update`,
-        );
+        )
       },
       nestedUpdatePatchType(options) {
-        const { table, constraint } = options;
+        const { table, constraint } = options
 
-        const tableFieldName = this.tableFieldName(table);
-        const parentTableFieldName = this.tableFieldName(constraint.class);
-        const constraintName = constraint.tags.name || constraint.name;
+        const tableFieldName = this.tableFieldName(table)
+        const parentTableFieldName = this.tableFieldName(constraint.class)
+        const constraintName = constraint.tags.name || constraint.name
 
         return this.camelCase(
           `update_${tableFieldName}_on_${parentTableFieldName}_for_${constraintName}_patch`,
-        );
+        )
       },
       nestedUpdateByKeyInputType(options) {
-        const { table, constraint, keyConstraint } = options;
+        const { table, constraint, keyConstraint } = options
 
-        const tableFieldName = this.tableFieldName(table);
-        const parentTableFieldName = this.tableFieldName(constraint.class);
-        const constraintName = constraint.tags.name || constraint.name;
-        const keyConstraintName = keyConstraint.tags.name || keyConstraint.name;
+        const tableFieldName = this.tableFieldName(table)
+        const parentTableFieldName = this.tableFieldName(constraint.class)
+        const constraintName = constraint.tags.name || constraint.name
+        const keyConstraintName = keyConstraint.tags.name || keyConstraint.name
 
         return this.upperCamelCase(
           `${tableFieldName}_on_${parentTableFieldName}_for_${constraintName}_using_${keyConstraintName}_update`,
-        );
+        )
       },
     }),
-  );
+  )
 
   builder.hook('build', (build) => {
     const {
@@ -62,7 +60,7 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
       pgOmit: omit,
       pgGetGqlTypeByTypeIdAndModifier,
       nodeIdFieldName,
-    } = build;
+    } = build
 
     return extend(build, {
       pgNestedTableUpdaterFields: {},
@@ -74,26 +72,21 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
         where,
         context,
       }) => {
-        const { foreignTable } = nestedField;
-        const { isNodeIdUpdater, constraint } = connectorField;
+        const { foreignTable } = nestedField
+        const { isNodeIdUpdater, constraint } = connectorField
 
-        let keyWhere = '';
+        let keyWhere = ''
 
         if (isNodeIdUpdater) {
-          const nodeId = input[nodeIdFieldName];
-          const primaryKeys = foreignTable.primaryKeyConstraint.keyAttributes;
-          const { Type, identifiers } = build.getTypeAndIdentifiersFromNodeId(
-            nodeId,
-          );
-          const ForeignTableType = pgGetGqlTypeByTypeIdAndModifier(
-            foreignTable.type.id,
-            null,
-          );
+          const nodeId = input[nodeIdFieldName]
+          const primaryKeys = foreignTable.primaryKeyConstraint.keyAttributes
+          const { Type, identifiers } = build.getTypeAndIdentifiersFromNodeId(nodeId)
+          const ForeignTableType = pgGetGqlTypeByTypeIdAndModifier(foreignTable.type.id, null)
           if (Type !== ForeignTableType) {
-            throw new Error('Mismatched type');
+            throw new Error('Mismatched type')
           }
           if (identifiers.length !== primaryKeys.length) {
-            throw new Error('Invalid ID');
+            throw new Error('Invalid ID')
           }
           keyWhere = sql.fragment`${sql.join(
             primaryKeys.map(
@@ -105,9 +98,9 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
                 )}`,
             ),
             ') and (',
-          )}`;
+          )}`
         } else {
-          const foreignPrimaryKeys = constraint.keyAttributes;
+          const foreignPrimaryKeys = constraint.keyAttributes
           keyWhere = sql.fragment`${sql.join(
             foreignPrimaryKeys.map(
               (k) => sql.fragment`
@@ -119,59 +112,47 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
               `,
             ),
             ') and (',
-          )}`;
+          )}`
         }
 
-        const patchField =
-          input[inflection.patchField(inflection.tableFieldName(foreignTable))];
-        const sqlColumns = [];
-        const sqlValues = [];
+        const patchField = input[inflection.patchField(inflection.tableFieldName(foreignTable))]
+        const sqlColumns = []
+        const sqlValues = []
         foreignTable.attributes.forEach((attr) => {
-          if (!pgColumnFilter(attr, build, context)) return;
-          if (omit(attr, 'update')) return;
+          if (!pgColumnFilter(attr, build, context)) return
+          if (omit(attr, 'update')) return
 
-          const colFieldName = inflection.column(attr);
+          const colFieldName = inflection.column(attr)
           if (colFieldName in patchField) {
-            const val = patchField[colFieldName];
-            sqlColumns.push(sql.identifier(attr.name));
-            sqlValues.push(gql2pg(val, attr.type, attr.typeModifier));
+            const val = patchField[colFieldName]
+            sqlColumns.push(sql.identifier(attr.name))
+            sqlValues.push(gql2pg(val, attr.type, attr.typeModifier))
           }
-        });
+        })
 
         if (sqlColumns.length === 0) {
           const selectQuery = sql.query`
             select *
-            from ${sql.identifier(
-              foreignTable.namespace.name,
-              foreignTable.name,
-            )}
-            where ${
-              where ? sql.fragment`(${keyWhere}) and (${where})` : keyWhere
-            }
-          `;
-          const { text, values } = sql.compile(selectQuery);
-          const { rows } = await pgClient.query(text, values);
-          return rows[0];
+            from ${sql.identifier(foreignTable.namespace.name, foreignTable.name)}
+            where ${where ? sql.fragment`(${keyWhere}) and (${where})` : keyWhere}
+          `
+          const { text, values } = sql.compile(selectQuery)
+          const { rows } = await pgClient.query(text, values)
+          return rows[0]
         }
 
         const updateQuery = sql.query`
-          update ${sql.identifier(
-            foreignTable.namespace.name,
-            foreignTable.name,
-          )}
-          set ${sql.join(
-            sqlColumns.map((col, i) => sql.fragment`${col} = ${sqlValues[i]}`),
-            ', ',
-          )}
+          update ${sql.identifier(foreignTable.namespace.name, foreignTable.name)}
+          set ${sql.join(sqlColumns.map((col, i) => sql.fragment`${col} = ${sqlValues[i]}`), ', ')}
           where ${where ? sql.fragment`(${keyWhere}) and (${where})` : keyWhere}
-          returning *`;
+          returning *`
 
-        const { text, values } = sql.compile(updateQuery);
-        const { rows } = await pgClient.query(text, values);
-        return rows[0];
+        const { text, values } = sql.compile(updateQuery)
+        const { rows } = await pgClient.query(text, values)
+        return rows[0]
       },
-    });
-  });
+    })
+  })
 
   builder.hook('GraphQLObjectType:fields', (fields, build, context) => {
     const {
@@ -186,38 +167,29 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
       pgOmit: omit,
       pgNestedTableUpdaterFields,
       graphql: { GraphQLNonNull, GraphQLInputObjectType, GraphQLID },
-    } = build;
+    } = build
     const {
       scope: { isRootMutation },
-    } = context;
+    } = context
 
     if (!isRootMutation) {
-      return fields;
+      return fields
     }
 
     introspectionResultsByKind.class
       .filter((cls) => cls.namespace && cls.isSelectable)
       .forEach((table) => {
-        if (
-          nestedMutationsTableNameWhiteList &&
-          !nestedMutationsTableNameWhiteList[table.name]
-        )
-          return;
-        pgNestedTableUpdaterFields[table.id] =
-          pgNestedTableUpdaterFields[table.id] || {};
+        if (nestedMutationsTableNameWhiteList && !nestedMutationsTableNameWhiteList[table.name])
+          return
+        pgNestedTableUpdaterFields[table.id] = pgNestedTableUpdaterFields[table.id] || {}
         introspectionResultsByKind.constraint
           .filter((con) => con.type === 'f')
-          .filter(
-            (con) =>
-              con.classId === table.id || con.foreignClassId === table.id,
-          )
+          .filter((con) => con.classId === table.id || con.foreignClassId === table.id)
           .filter((con) => !omit(con, 'read'))
           .filter((con) => !con.keyAttributes.some((key) => omit(key, 'read')))
           .forEach((constraint) => {
             const foreignTable =
-              constraint.classId === table.id
-                ? constraint.foreignClass
-                : constraint.class;
+              constraint.classId === table.id ? constraint.foreignClass : constraint.class
 
             /**
              * whitelist
@@ -227,19 +199,16 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
               nestedMutationsTableNameWhiteList[table.name] &&
               !nestedMutationsTableNameWhiteList[table.name][foreignTable.name]
             )
-              return;
+              return
 
-            const ForeignTableType = pgGetGqlTypeByTypeIdAndModifier(
-              foreignTable.type.id,
-              null,
-            );
-            const foreignTableFieldName = inflection.tableFieldName(
-              foreignTable,
-            );
-            const patchFieldName = inflection.patchField(foreignTableFieldName);
-            const ForeignTablePatch = getTypeByName(
-              inflection.patchType(ForeignTableType.name),
-            );
+            const ForeignTableType = pgGetGqlTypeByTypeIdAndModifier(foreignTable.type.id, null)
+            const foreignTableFieldName = inflection.tableFieldName(foreignTable)
+            const patchFieldName = inflection.patchField(foreignTableFieldName)
+            const ForeignTablePatch = getTypeByName(inflection.patchType(ForeignTableType.name))
+
+            if (!ForeignTablePatch) {
+              return
+            }
 
             const patchType = newWithHooks(
               GraphQLInputObjectType,
@@ -250,15 +219,11 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
                 }),
                 description: `An object where the defined keys will be set on the \`${foreignTableFieldName}\` being updated.`,
                 fields: () => {
-                  const omittedFields = constraint.keyAttributes.map((k) =>
-                    inflection.column(k),
-                  );
+                  const omittedFields = constraint.keyAttributes.map((k) => inflection.column(k))
                   return Object.keys(ForeignTablePatch._fields)
                     .filter((key) => !omittedFields.includes(key))
-                    .map((k) =>
-                      Object.assign({}, { [k]: ForeignTablePatch._fields[k] }),
-                    )
-                    .reduce((res, o) => Object.assign(res, o), {});
+                    .map((k) => Object.assign({}, { [k]: ForeignTablePatch._fields[k] }))
+                    .reduce((res, o) => Object.assign(res, o), {})
                 },
               },
               {
@@ -266,16 +231,14 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
                 pgInflection: foreignTable,
                 pgFieldInflection: constraint,
               },
-            );
+            )
 
             const foreignFields = foreignTable.constraints
               .filter((con) => con.type === 'u' || con.type === 'p')
               .filter((con) => !omit(con))
-              .filter(
-                (con) => !con.keyAttributes.some((key) => omit(key, 'read')),
-              )
+              .filter((con) => !con.keyAttributes.some((key) => omit(key, 'read')))
               .map((keyConstraint) => {
-                const keys = keyConstraint.keyAttributes;
+                const keys = keyConstraint.keyAttributes
 
                 // istanbul ignore next
                 if (!keys.every((_) => _)) {
@@ -283,7 +246,7 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
                     `Consistency error: could not find an attribute in the constraint when building nested connection type for ${describePgEntity(
                       foreignTable,
                     )}!`,
-                  );
+                  )
                 }
 
                 return {
@@ -339,10 +302,10 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
                       pgFieldInflection: constraint,
                     },
                   ),
-                };
-              });
+                }
+              })
 
-            const { primaryKeyConstraint: foreignPrimaryKey } = foreignTable;
+            const { primaryKeyConstraint: foreignPrimaryKey } = foreignTable
             if (nodeIdFieldName && foreignPrimaryKey) {
               foreignFields.push({
                 constraint: null,
@@ -356,8 +319,7 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
                       table,
                       constraint,
                     }),
-                    description:
-                      'The globally unique `ID` look up for the row to update.',
+                    description: 'The globally unique `ID` look up for the row to update.',
                     fields: {
                       [nodeIdFieldName]: {
                         description: `The globally unique \`ID\` which identifies a single \`${foreignTableFieldName}\` to be connected.`,
@@ -376,13 +338,13 @@ module.exports = function PostGraphileNestedUpdatersPlugin(
                     pgInflection: foreignTable,
                   },
                 ),
-              });
+              })
             }
 
-            pgNestedTableUpdaterFields[table.id][constraint.id] = foreignFields;
-          });
-      });
+            pgNestedTableUpdaterFields[table.id][constraint.id] = foreignFields
+          })
+      })
 
-    return fields;
-  });
-};
+    return fields
+  })
+}