+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changes
 
+## v1.0.1
+
+  * Correctly release the savepoint on error (thanks @sijad).
+  
+## v1.0.0
+
+  * Bump dependencies.
+  * Guard against creating updater fields where constraint is not available.
+
 ## v1.0.0-alpha.22
 
   * Fix bug where if an update mutation was called that did not locate